+++ src/Core/src/Platform/iOS/MauiPicker.cs
@@ -1,0 +1,25 @@
+﻿using System.Collections.Generic;
+using Foundation;
+using UIKit;
+using ObjCRuntime;
+
+namespace Microsoft.Maui
+{
+	public class MauiPicker : NoCaretField
+	{
+		readonly HashSet<string> _enableActions;
+
+		public MauiPicker(UIPickerView? uIPickerView)
+		{
+			UIPickerView = uIPickerView;
+
+			string[] actions = { "copy:", "select:", "selectAll:" };
+			_enableActions = new HashSet<string>(actions);
+		}
+
+		public UIPickerView? UIPickerView { get; set; }
+
+		public override bool CanPerform(Selector action, NSObject? withSender)
+			=> _enableActions.Contains(action.Name);
+	}
+}
