+++ src/Core/src/Platform/iOS/NoCaretField.cs
@@ -1,0 +1,20 @@
+﻿using UIKit;
+using RectangleF = CoreGraphics.CGRect;
+
+namespace Microsoft.Maui
+{
+	public class NoCaretField : UITextField
+	{
+		public NoCaretField() : base(new RectangleF())
+		{
+			SpellCheckingType = UITextSpellCheckingType.No;
+			AutocorrectionType = UITextAutocorrectionType.No;
+			AutocapitalizationType = UITextAutocapitalizationType.None;
+		}
+
+		public override RectangleF GetCaretRectForPosition(UITextPosition? position)
+		{
+			return new RectangleF();
+		}
+	}
+}
