+++ src/Core/src/Platform/iOS/PickerExtensions.cs
@@ -7,13 +7,13 @@ namespace Microsoft.Maui
 {
 	public static class PickerExtensions
 	{
-		public static void UpdateTitle(this NativePicker nativePicker, IPicker picker) =>
+		public static void UpdateTitle(this MauiPicker nativePicker, IPicker picker) =>
 			nativePicker.UpdatePicker(picker);
 				
-		public static void UpdateSelectedIndex(this NativePicker nativePicker, IPicker picker) =>
+		public static void UpdateSelectedIndex(this MauiPicker nativePicker, IPicker picker) =>
 			nativePicker.SetSelectedIndex(picker, picker.SelectedIndex);
 
-		internal static void UpdatePicker(this NativePicker nativePicker, IPicker picker)
+		internal static void UpdatePicker(this MauiPicker nativePicker, IPicker picker)
 		{
 			var selectedIndex = picker.SelectedIndex;
 			var items = picker.Items;
@@ -30,7 +30,7 @@ namespace Microsoft.Maui
 			nativePicker.SetSelectedItem(picker);
 		}
 
-		internal static void SetSelectedIndex(this NativePicker nativePicker, IPicker picker, int selectedIndex = 0)
+		internal static void SetSelectedIndex(this MauiPicker nativePicker, IPicker picker, int selectedIndex = 0)
 		{
 			picker.SelectedIndex = selectedIndex;
 
@@ -45,7 +45,7 @@ namespace Microsoft.Maui
 			pickerView?.Select(Math.Max(selectedIndex, 0), 0, true);
 		}
 
-		internal static void SetSelectedItem(this NativePicker nativePicker, IPicker picker)
+		internal static void SetSelectedItem(this MauiPicker nativePicker, IPicker picker)
 		{
 			if (nativePicker == null)
 				return;
@@ -67,7 +67,7 @@ namespace Microsoft.Maui
 			picker.SelectedItem = picker.Items[index];
 		}
 
-		internal static void UpdateAttributedPlaceholder(this NativePicker nativePicker, NSAttributedString nsAttributedString)
+		internal static void UpdateAttributedPlaceholder(this MauiPicker nativePicker, NSAttributedString nsAttributedString)
 		{
 			nativePicker.AttributedPlaceholder = nsAttributedString;
 		}
